Deprecate `Bot.subscribe*` for better Coroutine life cycle management; Add docs
